+++ src/main/java/com/reason/ide/files/MlgFile.java
@@ -3,23 +3,21 @@ package com.reason.ide.files;
 import com.intellij.extapi.psi.*;
 import com.intellij.openapi.fileTypes.FileType;
 import com.intellij.psi.FileViewProvider;
-import com.reason.lang.extra.OclMlgLanguage;
+import com.reason.lang.ocamlgrammar.*;
 import org.jetbrains.annotations.NotNull;
 
 public class MlgFile extends PsiFileBase {
-  public MlgFile(@NotNull FileViewProvider viewProvider) {
-    super(viewProvider, OclMlgLanguage.INSTANCE);
-  }
+    public MlgFile(@NotNull FileViewProvider viewProvider) {
+        super(viewProvider, OclGrammarLanguage.INSTANCE);
+    }
 
-  @NotNull
-  @Override
-  public FileType getFileType() {
-    return MlgFileType.INSTANCE;
-  }
+    @Override
+    public @NotNull FileType getFileType() {
+        return MlgFileType.INSTANCE;
+    }
 
-  @NotNull
-  @Override
-  public String toString() {
-    return MlgFileType.INSTANCE.getDescription();
-  }
+    @Override
+    public @NotNull String toString() {
+        return MlgFileType.INSTANCE.getDescription();
+    }
 }