+++ src/main/java/com/reason/ide/files/MlgFileType.java
@@ -1,40 +1,40 @@
 package com.reason.ide.files;
 
-import com.intellij.openapi.fileTypes.LanguageFileType;
-import com.reason.lang.extra.OclMlgLanguage;
+import com.intellij.openapi.fileTypes.*;
+import com.reason.ide.highlight.*;
 import com.reason.ide.ORIcons;
+
 import javax.swing.*;
+
+import com.reason.lang.ocamlgrammar.*;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class MlgFileType extends LanguageFileType {
-  public static final MlgFileType INSTANCE = new MlgFileType();
+    public static final MlgFileType INSTANCE = new MlgFileType();
 
-  private MlgFileType() {
-    super(OclMlgLanguage.INSTANCE);
-  }
+    private MlgFileType() {
+        super(OclGrammarLanguage.INSTANCE);
+        FileTypeEditorHighlighterProviders.INSTANCE.addExplicitExtension(this, (project, fileType, virtualFile, colors) -> new OclGrammarEditorHighlighter(project, virtualFile, colors));
+    }
 
-  @NotNull
-  @Override
-  public String getName() {
-    return "MLG";
-  }
+    @Override
+    public @NotNull String getName() {
+        return "MLG";
+    }
 
-  @NotNull
-  @Override
-  public String getDescription() {
-    return "OCaml grammar file";
-  }
+    @Override
+    public @NotNull String getDescription() {
+        return "OCaml grammar file";
+    }
 
-  @NotNull
-  @Override
-  public String getDefaultExtension() {
-    return "mlg";
-  }
+    @Override
+    public @NotNull String getDefaultExtension() {
+        return "mlg";
+    }
 
-  @Nullable
-  @Override
-  public Icon getIcon() {
-    return ORIcons.OCL_GREEN_FILE;
-  }
+    @Override
+    public @Nullable Icon getIcon() {
+        return ORIcons.OCL_GREEN_FILE;
+    }
 }