+++ src/main/java/com/reason/ide/structure/ORStructureViewModel.java
@@ -31,6 +31,7 @@ public class ORStructureViewModel extends StructureViewModelBase implements com.
                 || element instanceof ResFile
                 || element instanceof OclFile
                 || element instanceof MlyFile
+                || element instanceof MlgFile
                 || element instanceof DuneFile;
     }
 }