+++ src/main/java/com/reason/lang/ocamllex/OclLexAstFactory.java
@@ -4,6 +4,7 @@ import com.intellij.extapi.psi.*;
 import com.intellij.lang.*;
 import com.intellij.psi.*;
 import com.intellij.psi.tree.*;
+import com.reason.lang.core.psi.*;
 import com.reason.lang.core.psi.ocamllex.*;
 import org.jetbrains.annotations.*;
 
@@ -24,7 +25,7 @@ class OclLexAstFactory extends ASTFactory {
             return new RPsiLexPattern(node);
         }
         if (type == OclLexTypes.INSTANCE.C_INJECTION) {
-            return new RPsiLexInjection(node);
+            return new RPsiOCamlInjection(node);
         }
 
         return new ASTWrapperPsiElement(node);