+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 > - :nail_care:  [Polish]
 
 (_Tags are copied from [babel](https://github.com/babel/babel/blob/master/CHANGELOG.md)_)
+                        
+## Unreleased
+
+- :rocket: ![o] [#118](https://github.com/giraud/reasonml-idea-plugin/issues/118) Much better parsing of mlg files (ocaml injection, highlighting, folding)
 
 ## 0.113 - 2023/05/29